enhancement: add changelog and simplify registry package
