+++ changelog/unreleased/client-selector-pool.md
@@ -1,0 +1,6 @@
+Enhancement: client selector pool
+
+Add the ability to use iterable client pools for the grpc service communication,
+the underlying grpc client and connection is fetched randomly from the available services.
+
+https://github.com/cs3org/reva/pull/3939
